backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
sqlalchemy
asyncpg
alembic
aiosqlite
python-jose
passlib
PyGithub
openai
anthropic
google-generativeai
python-multipart
email-validator
cryptography
sse-starlette
mistralai-gcp
